few refactors
