+++ src/app/auth/redirect/redirect.component.ts
@@ -24,11 +24,8 @@ export class RedirectComponent implements OnInit {
     if (this.route.snapshot.queryParams?.state === 'travelperk_local_redirect') {
       const url = `http://localhost:4200/integrations/travelperk?code=${this.route.snapshot.queryParams.code}`;
       this.windowService.redirect(url);
-    } else if (this.route.snapshot.queryParams?.state === 'qbo_local_redirect') {
-      const url = `http://localhost:4200/integrations/qbo/oauth_callback?code=${this.route.snapshot.queryParams.code}&realmId=${this.route.snapshot.queryParams.realmId}`;
-      this.windowService.redirect(url);
-    } else if (this.route.snapshot.queryParams?.state === 'business_central_local_redirect') {
-      const url = `http://localhost:4200/integrations/business_central/onboarding/connector?code=${this.route.snapshot.queryParams.code}`;
+    } else if (this.route.snapshot.queryParams?.state === 'qbo_local_redirect' || this.route.snapshot.queryParams?.state === 'business_central_local_redirect') {
+      const url = `http://localhost:4200/integrations/oauth_callback?code=${this.route.snapshot.queryParams.code}&realmId=${this.route.snapshot.queryParams.realmId}`;
       this.windowService.redirect(url);
     } else {
       this.redirectToFyleOAuth();
