+++ src/app/integrations/oauth-callback/oauth-callback.component.html
@@ -1,0 +1,3 @@
+<div class="tw-flex tw-flex-col tw-justify-center tw-items-center tw-h-[100%]">
+    <app-loader></app-loader>
+</div>
