+++ src/app/integrations/oauth-callback/oauth-callback.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { OauthCallbackComponent } from './oauth-callback.component';
+
+describe('OauthCallbackComponent', () => {
+  let component: OauthCallbackComponent;
+  let fixture: ComponentFixture<OauthCallbackComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ OauthCallbackComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(OauthCallbackComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
