+++ src/app/integrations/oauth-callback/oauth-callback.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-oauth-callback',
+  templateUrl: './oauth-callback.component.html',
+  styleUrls: ['./oauth-callback.component.scss']
+})
+export class OauthCallbackComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
