+++ src/app/integrations/qbo/qbo-onboarding/qbo-onboarding-connector/qbo-onboarding-connector.component.html
@@ -22,7 +22,8 @@
                     [headerText]="'Incorrect account selected'"
                     [contextText]="'You had previously set up the integration with a different QuickBooks Online account. Please choose the same to restore the settings'"
                     [iconPath]="'assets/icons/alert_dialog.svg'"
-                    [confirmBtnText]="'Re connect'">
+                    [confirmBtnText]="'Re connect'"
+                    [showSecondaryCTA]="false">
                 </app-configuration-confirmation-dialog>
             </div>
             <div>
