+++ src/app/integrations/qbo/qbo-onboarding/qbo-onboarding-landing/qbo-onboarding-landing.component.ts
@@ -1,5 +1,14 @@
 import { Component, OnInit } from '@angular/core';
+import { NavigationExtras, Router } from '@angular/router';
 import { brandingConfig, brandingDemoVideoLinks, brandingKbArticles } from 'src/app/branding/branding-config';
+import { QBOOnboardingState, ToastSeverity } from 'src/app/core/models/enum/enum.model';
+import { QBOCredential } from 'src/app/core/models/qbo/db/qbo-credential.model';
+import { QBOConnectorModel, QBOConnectorPost } from 'src/app/core/models/qbo/qbo-configuration/qbo-connector.model';
+import { HelperService } from 'src/app/core/services/common/helper.service';
+import { IntegrationsToastService } from 'src/app/core/services/common/integrations-toast.service';
+import { WorkspaceService } from 'src/app/core/services/common/workspace.service';
+import { QboConnectorService } from 'src/app/core/services/qbo/qbo-configuration/qbo-connector.service';
+import { environment } from 'src/environments/environment';
 
 @Component({
   selector: 'app-qbo-onboarding-landing',
@@ -14,7 +23,66 @@ export class QboOnboardingLandingComponent implements OnInit {
 
   embedVideoLink = brandingDemoVideoLinks.onboarding.QBO;
 
-  constructor() { }
+  isIncorrectQBOConnectedDialogVisible: boolean = false;
+
+  constructor(
+    private helperService: HelperService,
+    private qboConnectorService: QboConnectorService,
+    private router: Router,
+    private toastService: IntegrationsToastService,
+    private workspaceService: WorkspaceService
+  ) { }
+
+  acceptWarning(isWarningAccepted: boolean): void {
+    this.isIncorrectQBOConnectedDialogVisible = false;
+    if (isWarningAccepted) {
+      this.router.navigate([`/integrations/qbo/onboarding/landing`]);
+    }
+  }
+
+  connectQbo(): void {
+    const url = `${environment.qbo_authorize_uri}?client_id=${environment.qbo_oauth_client_id}&scope=com.intuit.quickbooks.accounting&response_type=code&redirect_uri=${environment.qbo_oauth_redirect_uri}&state=qbo_local_redirect`;
+
+    this.helperService.oauthCallbackUrl.subscribe((callbackURL: string) => {
+      const code = callbackURL.split('code=')[1].split('&')[0];
+      const realmId = callbackURL.split('realmId=')[1].split('&')[0];
+      this.checkProgressAndRedirect(code, realmId);
+    });
+
+    this.helperService.oauthHandler(url);
+  }
+
+  private postQboCredentials(code: string, realmId: string): void {
+    const payload: QBOConnectorPost = QBOConnectorModel.constructPayload(code, realmId);
+
+    this.qboConnectorService.connectQBO(payload).subscribe((qboCredential: QBOCredential) => {
+      this.router.navigate([`/integrations/qbo/main/dashboard`]);
+    }, (error) => {
+      const errorMessage = 'message' in error.error ? error.error.message : 'Failed to connect to QuickBooks Online. Please try again';
+      if (errorMessage === 'Please choose the correct QuickBooks Online account') {
+        this.isIncorrectQBOConnectedDialogVisible = true;
+      } else {
+        this.toastService.displayToastMessage(ToastSeverity.ERROR, errorMessage);
+        this.router.navigate([`/integration/qbo/onboarding/landing`]);
+      }
+    });
+  }
+
+  private checkProgressAndRedirect(code: string, realmId: string): void {
+    const onboardingState: QBOOnboardingState = this.workspaceService.getOnboardingState();
+    const navigationExtras: NavigationExtras = {
+      queryParams: {
+        code,
+        realmId
+      }
+    };
+
+    if (onboardingState !== QBOOnboardingState.COMPLETE) {
+      this.router.navigate(['integrations/qbo/onboarding/connector'], navigationExtras);
+    } else {
+      this.postQboCredentials(code, realmId);
+    }
+  }
 
   ngOnInit(): void {
   }
